+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "dashboard",
-  "version": "2.4.0",
+  "version": "2.5.0",
   "private": true,
   "dependencies": {
     "@chakra-ui/icons": "^2.1.1",
